requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
DateTime==4.4
Django==4.0.4
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-form==3.4
django-bootstrap3==21.2
django-bootstrap4==22.1
django-filter==21.1
Pillow==9.1.0
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
zope.interface==5.4.0
